Switch to using GCS for images [Work in Progress]
This will only work for PNG images, but that  will be okay since the image resize tool will always output images as PNGs. This is not cone until the image resize/rotate tool is done since it depends on the PNG file type. See an example here:  https://hacks.mozilla.org/2011/01/how-to-develop-a-html5-image-uploader/